refactor: Extension 코드 별도의 파일로 분리
