+++ BanchanApp/BanchanApp/Controllers/BanchanDetailViewController.swift
@@ -109,6 +109,7 @@ class BanchanDetailViewController: UIViewController {
     // MARK: - Life Cycle
     override func viewDidLoad() {
         super.viewDidLoad()
+		self.view.backgroundColor = .white
 
         self.configureUI()
         self.carouselView.delegate = self
