+++ BanchanApp/BanchanApp/Views/ProductDetailView.swift
@@ -75,56 +75,3 @@ class ProductDetailView: UIView {
     }
 
 }
-
-extension UILabel {
-
-    func applyStrikethoroughStyle() {
-        if let text = self.text {
-            let attributeString = NSMutableAttributedString(string: text)
-            attributeString.addAttribute(.strikethroughStyle, value: NSUnderlineStyle.single.rawValue, range: NSRange(location: 0, length: attributeString.length))
-            self.attributedText = attributeString
-        }
-    }
-
-    func applySemiBold() {
-        self.font = .systemFont(ofSize: 14, weight: .semibold)
-    }
-
-    func applyColorGray() {
-        self.font = .systemFont(ofSize: 14)
-        self.textColor = .systemGray2
-    }
-
-}
-
-class BadgeLabel: UILabel {
-
-    private var padding = UIEdgeInsets(top: 4, left: 16, bottom: 4, right: 16)
-
-    convenience init(color: UIColor) {
-        self.init()
-        setupStyle(with: color)
-    }
-
-    private func setupStyle(with color: UIColor) {
-        self.font = .systemFont(ofSize: 12, weight: .semibold)
-        self.textColor = .white
-        self.backgroundColor = color
-        self.textAlignment = .center
-        self.layer.masksToBounds = true
-        self.layer.cornerCurve = .continuous
-        self.layer.cornerRadius = 15
-    }
-
-    override func drawText(in rect: CGRect) {
-        super.drawText(in: rect.inset(by: padding))
-    }
-
-    override var intrinsicContentSize: CGSize {
-        var contentSize = super.intrinsicContentSize
-        contentSize.height += padding.top + padding.bottom
-        contentSize.width += padding.left + padding.right
-        return contentSize
-    }
-
-}
